requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Uploads==0.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.27
alembic==1.13.1
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
python-dateutil==2.8.2
pytz==2024.1
flask-fontawesome
google-generativeai==0.6.0
beautifulsoup4==4.12.2
razorpay
setuptools
flask-dance[sqla]